requirements.txt
flask
uwsgi
jinja2
requests
python-dotenv
tqdm
pandas